[refactor] refactoring codes as below
1) split class files
2) add common data files to manage common constants
