+++ hardware/ssd_common_data.py
@@ -1,0 +1,11 @@
+CMD_READ_TYPE = 'R'
+CMD_WRITE_TYPE = 'W'
+CMD_ERASE_TYPE = 'E'
+CMD_FLUSH_TYPE = 'F'
+
+MAX_ERASE_SIZE = 10
+INITIAL_DATA_VALUE = '0x00000000'
+
+MAX_DATA_LEN = 100
+MAX_RESULT_LEN = 1
+MAX_BUFFER_LEN = 10
